+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Types of changes
 - Added a warning for entities that are ignored/skipped due to the maximum entity length or maximum model input length.
 - Added info-level logs displaying the detected labeling scheme (IOB/IOB2, BIOES, BILOU, none).
 - Added a warning suggesting to use `model.cuda()` when predictions are performed on a CPU while CUDA is available.
+- Added `try_cuda` method to `SpanMarkerModel` which tries to place the model on CUDA and does nothing if that fails.
 
 ### Changed